requirements.txt
requests
tqdm
pycryptodome
urllib3
gzip
